LayoutTests/media/media-can-play-ogg-expected.txt

Test HTMLMediaElement canPlayType() method for ogg media containers.

EXPECTED (video.canPlayType('Audio/ogg') == 'maybe') OK
EXPECTED (video.canPlayType('Video/ogg') == 'maybe') OK
END OF TEST
